add metaprogramming magic to patches
