+++ lib/vizsla/subscribers.rb
@@ -10,6 +10,8 @@ module Vizsla
       collect_events_data
     end
 
+    private
+
     def collect_events_data
       if rails_app?
         rails_hooks
@@ -34,29 +36,24 @@ module Vizsla
       resque_hook
     end
 
+    # ===---------------------------===
+    # Rails Hooks
+    # ===---------------------------===
+
     def sql_hook
-      ActiveSupport::Notifications.subscribe "sql.active_record" do |*args|
-        event = SQLEvent.new(args)
-        @events_data << event if event.valid?
-      end
+      subscribe_asn 'sql.active_record', SQLEvent
     end
 
     def process_action_hook
-      ActiveSupport::Notifications.subscribe "process_action.action_controller" do |*args|
-        event = ControllerEvent.new(args)
-        @events_data << event
-      end
+      subscribe_asn 'process_action.action_controller', ControllerEvent
     end
 
     def render_template_hook
-      ActiveSupport::Notifications.subscribe "render_template.action_view" do |*args|
-        event = ViewEvent.new(args)
-        @events_data << event
-      end
+      subscribe_asn 'render_template.action_view', ViewEvent
     end
 
     # ===---------------------------===
-    # Non-Rails Hooks
+    # DB Hooks
     # ===---------------------------===
 
     def postgres_hook
@@ -76,6 +73,10 @@ module Vizsla
       end
     end
 
+    # ===---------------------------===
+    # Background Job Hooks
+    # ===---------------------------===
+
     def sidekiq_hook
       return unless defined? ::Sidekiq
       ::Vizsla::BackgroundJobInstrumentation.install :sidekiq
@@ -91,14 +92,22 @@ module Vizsla
     # ===---------------------------===
 
     def sinatra_hook
-      return unless sinatra_app?
       ::Vizsla::Patches.patch_sinatra do |event_data|
         event = SinatraEvent.new event_data
         @events_data << event
       end
     end
 
-    private
+    # ===---------------------------===
+    # Aux
+    # ===---------------------------===
+
+    def subscribe_asn(event_name, event_klass)
+      ActiveSupport::Notifications.subscribe event_name do |*args|
+        event = event_klass.new args
+        @events_data << event if event.valid?
+      end
+    end
 
     def rails_app?
       defined? ::Rails
